Added constrained bodies and Overleaf link
This was a huge update. The old versions use springs or questionable methods (simulating each bond individually) to simulate rotating bodies, but I went all in and implemented them properly with angular momentum/torques because the old simulations were noticeably innacurate (ODE drift I guess). this.pos was replaced with this.cm throughout in keeping with these changes.